LaraTrustAndCustomerControllerUpdate

+++ app/Http/Controllers/Api/CustomerController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Api;
 use App\Http\Controllers\Controller;
 use App\Http\Resources\CustomerResource;
 use App\Models\Customer;
+use App\Policies\CustomerPolicy;
 use Illuminate\Http\Request;
 
 class CustomerController extends Controller
@@ -16,6 +17,7 @@ class CustomerController extends Controller
      */
     public function index()
     {
+        $this->authorize('index', CustomerPolicy::class);
         return CustomerResource::collection(Customer::all());
     }
 
@@ -36,6 +38,7 @@ class CustomerController extends Controller
      */
     public function store(Request $request)
     {
+        $this->authorize('store', CustomerPolicy::class);
         $customer = Customer::create($request->all());
         return response()->json($customer, 201);
     }
@@ -48,6 +51,7 @@ class CustomerController extends Controller
      */
     public function show(Customer $customer)
     {
+        $this->authorize('show', CustomerPolicy::class);
         return response()->json($customer, 200);
     }
 
@@ -59,7 +63,7 @@ class CustomerController extends Controller
      */
     public function edit($id)
     {
-        //
+        $this->authorize('edit', CustomerPolicy::class);
     }
 
     /**
@@ -71,6 +75,7 @@ class CustomerController extends Controller
      */
     public function update(Request $request, Customer $customer)
     {
+        $this->authorize('update', CustomerPolicy::class);
         $customer->update($request->all());
         return response()->json($customer, 200);
     }
@@ -83,6 +88,7 @@ class CustomerController extends Controller
      */
     public function destroy(Customer $customer)
     {
+        $this->authorize('destroy', CustomerPolicy::class);
         $customer->delete();
         return response()->json(NULL, 204);
     }
